% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{estimation.VCM}
\alias{estimation.VCM}
\title{Function to calculate parameter estimations from many MCMC chains generated by the \code{VCM_fct} function}
\usage{
estimation.VCM(object)
}
\arguments{
\item{object}{list generated by the \code{VCM_fct} function}
}
\value{
returne the input object including the parameter estimations
}
\description{
Function to calculate parameter estimations from many MCMC chains generated by the \code{VCM_fct} function
}
