% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{gelman.diag.VCM}
\alias{gelman.diag.VCM}
\title{Gelman-Rubin's convergence diagnostics of the MCMC chain generated by the \code{VCM_fct} function}
\usage{
gelman.diag.VCM(object, gelman.plot = FALSE, traceplot = FALSE)
}
\arguments{
\item{object}{list generated by the \code{VCM_fct} function}

\item{gelman.plot}{boolean indicating if the gelman plot should be plotted}

\item{traceplot}{boolean indicating if the trace plot should be plotted}
}
\value{
return input \code{object} including Gelman Rubin's convergence diagnostics
}
\description{
This function can be apply only if the number of repetition applied in the \code{VCM_fct} function is upper than one.
}
