% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{plot_diagnostic_gamma}
\alias{plot_diagnostic_gamma}
\title{Function to plot visual convergence diagnostic of variable marginal posterior inclusion probabilities}
\usage{
plot_diagnostic_gamma(object)
}
\arguments{
\item{object}{list generated by the \code{VCM_fct} function}
}
\value{

}
\description{
Function to plot visual convergence diagnostic of variable marginal posterior inclusion probabilities
}
