% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{plot_functional_effects}
\alias{plot_functional_effects}
\title{Function to plot the estimated functional effects}
\usage{
plot_functional_effects(
  object,
  mfrow = c(6, 7),
  mar = c(1, 1, 4, 1),
  plot = c("Y", "mu", "env", "beta"),
  id = which(object$estimation$mean.marginal.probabilities > 0.5),
  add = c("quantile", "matplot"),
  name = NULL
)
}
\arguments{
\item{object}{list generated by the \code{VCM_fct} function}

\item{mfrow}{a numerical vector of the form c(nr, nc). Subsequent figures will be drawn in an nr-by-nc array on the device by rows. The default is c(6, 7).}

\item{mar}{a numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. The default is c(1, 1, 4, 1).}

\item{plot}{caracter list ("Y", "mu", "env", "beta") containing object names for which you want the plot curves. The default is ("Y", "mu", "env", "beta").}

\item{id}{variable indinces for which you want to plot the functional effects. By defaults the argument id contains indicies of variables with posterior marginal probabilities upper than 0.5.}

\item{add}{caracter list ("quantile", "matplot") to indicate if you want to add credible intervals at 95\% or matpot of estimated curve of each repetition.}

\item{name}{optional caracter vector of variable names}
}
\value{

}
\description{
Function to plot the estimated functional effects
}
\examples{
R/exemple_1.R
}
