\name{VCGSS-package}
\alias{VCGSS-package}
\alias{VCGSS}
\docType{package}
\title{
\packageTitle{VCGSS}
}
\description{
\packageDescription{VCGSS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VCGSS}
\packageIndices{VCGSS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VCGSS}

Maintainer: \packageMaintainer{VCGSS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
