% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{VCM_fct}
\alias{VCM_fct}
\title{Main function to run the Bayesian varying coefficient model using spike-and-slab prior for variable selection}
\usage{
VCM_fct(
  Y,
  X,
  ENV = NULL,
  selection = TRUE,
  interpolation = "P-spline",
  df = floor(ncol(Y)/3) + 3,
  order_diff = 2,
  niter = 10000,
  burnin = 5000,
  thin = 10,
  rep = 1,
  cores = -1,
  save = TRUE,
  path = NULL,
  summary.fct = mean,
  scale = FALSE,
  estimation = TRUE,
  gelman.diag = TRUE
)
}
\arguments{
\item{Y}{n x T matrix of phenotypic observations (n is the number of individuals, T is the number of observed time points)}

\item{X}{n x q matrix of variables (markers), q is the number of variables}

\item{ENV}{n x ne matrix of environmental variables, ne is the number of environmental variables}

\item{selection}{boolean indicating if selection is requiered, the default is TRUE}

\item{interpolation}{character, indicating the approach used to estimate curve of the effects. Possible choise is c('Legendre', 'B-spline', 'P-spline', 'RW'), the default is 'P-spline'.}

\item{df}{integer, only for Legendre, B-spline or P-spline approach. df is the degree of freedom parameter for these approaches, the default is floor(ncol(Y)/3)+3}

\item{order_diff}{interger in [1, 2, 3], order of the difference penalty, the default is 2.}

\item{niter}{number of iterations of the MCMC chain, the default is 10000}

\item{burnin}{number of initial iterations of the MCMC chain which should be discarded, the defaults to 5000}

\item{thin}{save only every thin-th iteration, the default is 10}

\item{rep}{number of repetitions, the default is one}

\item{cores}{number of cores for parallelisation, the default is -1 indicating that all cores minus 1 are used.}

\item{save}{boolean indicating if the MCMC chains should me saved, the default is TRUE}

\item{path}{optional (existing) path to save the MCMC chains if save = TRUE, if NULL, path = Fused_HS if selection = TRUE, Fusion_HS ortherwise}

\item{summary.fct}{name of the function (mean, mode or median) used to calculate the estimation of the parameters from the MCMC posterior samples, the default is mean.}

\item{scale}{boolean, indicating if X should be scaled, the default is TRUE}

\item{estimation}{boolean, indicating summary of MCMC chains should be apply to provide estimation of parameters and curve, the default is TRUE}

\item{gelman.diag}{boolean, indicating gelman.diag should be apply on MCMC chains, the default is TRUE}

\item{...}{additional arguments}
}
\value{
Return a list
}
\description{
Main function to run the Bayesian varying coefficient model using spike-and-slab prior for variable selection
}
