% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_VCM.R
\name{gelman.diag.VCM}
\alias{gelman.diag.VCM}
\title{Gelman Rubin's convergence diagnostic of the MCMC chain generated by the \code{VCM_fct} function}
\usage{
gelman.diag.VCM(object, gelman.plot = FALSE, traceplot = FALSE)
}
\arguments{
\item{object}{list generated by the \code{VCM_fct} function}

\item{gelman.plot}{boolean indicating if the gelman plot should be plotted in the case where K is upper than one, used only if gelman.diag = TRUE, the default is FALSE}

\item{traceplot}{boolean indicating if the traceplot should be plotted in the case where K is upper than one, used only if gelman.diag = TRUE, the default is FALSE}
}
\value{
return input object including Gelman Rubin's convergence diagnostic
}
\description{
Gelman Rubin's convergence diagnostic of the MCMC chain generated by the \code{VCM_fct} function
}
